opendota_supabase/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
supabase>=2.4.0
pandas>=2.2.0
streamlit>=1.32.0
